Rename an interface and move it to another namespace
